requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
altair==4.0.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
bayesian-optimization==3.0.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.57.0
datasets==3.6.0
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.7.0
docker==7.1.0
editdistpy==0.1.5
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keybert==0.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
looseversion==1.3.0
lxml==6.0.0
macholib==1.16.3
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlflow==3.1.0
mlflow-skinny==3.1.0
mpmath==1.3.0
multidict==6.6.2
multiprocess==0.70.16
murmurhash==1.0.13
mysql-connector-python==9.2.0
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nlpk==0.0.6
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.2
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.3.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==3.20.3
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
Pympler==1.1
PyMuPDF==1.26.1
pyparsing==3.2.3
PyPDF2==3.0.1
pyqt-toast-notification==1.3.3
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pyspellchecker==0.8.2
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.3.0
QtPy==2.4.3
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==78.1.0
shap==0.48.0
shellingham==1.5.4
shiboken6==6.8.2.1
simplejson==3.20.1
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.16.0
sympy==1.14.0
symspellpy==6.9.0
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.51.3
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.21.0
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==3.0.0
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.3.31
zipp==3.21.0
